// WADNR.Web/src/app/shared/generated/model/create-workflow-progress-response.ts
export class CreateWorkflowProgressResponse { 
    ProjectID?: number;
    ProjectName?: string | null;
    ProjectApprovalStatusID?: number;
    ProjectApprovalStatusName?: string | null;
    CanSubmit?: boolean;
    CreatedByPersonName?: string | null;
    CreatedByOrganizationName?: string | null;
    CreateDate?: string | null;
    Steps?: CreateWorkflowProgressResponseSteps | null;
    CanApprove?: boolean;
    CanReject?: boolean;
    CanReturn?: boolean;
    CanWithdraw?: boolean;
    CanEdit?: boolean;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}